melengkapi method : user dengan akses admin tidak boleh didelete

+++ app/models/User.php
@@ -25,9 +25,9 @@ class User {
 
     public function get_user($level = FALSE) {
 
-        $sql = "SELECT ".$this->_table.".KD_USER, ".$this->_table.".NIP_USER, ".$this->_table.".NM_USER, ".$this->_table.".PASS_USER,
-                ".$this->_table2.".nama as AKSES_USER, ".$this->_table.".FOTO_USER FROM ".$this->_table." LEFT JOIN ".$this->_table2."
-                on ".$this->_table.".AKSES_USER = ".$this->_table2.".id ";
+        $sql = "SELECT " . $this->_table . ".KD_USER, " . $this->_table . ".NIP_USER, " . $this->_table . ".NM_USER, " . $this->_table . ".PASS_USER,
+                " . $this->_table2 . ".nama as AKSES_USER, " . $this->_table . ".FOTO_USER FROM " . $this->_table . " LEFT JOIN " . $this->_table2 . "
+                on " . $this->_table . ".AKSES_USER = " . $this->_table2 . ".id ";
         if ($level) {
             $sql .=" WHERE AKSES_USER= '2'";
         }
@@ -101,38 +101,38 @@ class User {
     }
 
     public function updateUser(User $user) {
-      
-            $where = "KD_USER = " . $user->get_id();
 
-            $data = array(
-                'NIP_USER' => $user->get_nip(),
-                'NM_USER' => $user->get_nmUser(),
-                'PASS_USER' => Hash::create('sha1', $user->get_pass(), HASH_SALT_KEY),
-                'AKSES_USER' => $user->get_akses(),
-                'FOTO_USER' => $user->get_foto()
-            );
-            $this->_db->update($this->_table, $data, $where);
-       
+        $where = "KD_USER = " . $user->get_id();
+
+        $data = array(
+            'NIP_USER' => $user->get_nip(),
+            'NM_USER' => $user->get_nmUser(),
+            'PASS_USER' => Hash::create('sha1', $user->get_pass(), HASH_SALT_KEY),
+            'AKSES_USER' => $user->get_akses(),
+            'FOTO_USER' => $user->get_foto()
+        );
+        $this->_db->update($this->_table, $data, $where);
     }
-    public function updateUser_withoutpass (User $user){
-            $where = "KD_USER = " . $user->get_id();
+
+    public function updateUser_withoutpass(User $user) {
+        $where = "KD_USER = " . $user->get_id();
 
 //            var_dump($user->get_pass());
-            $data = array(
-                'NIP_USER' => $user->get_nip(),
-                'NM_USER' => $user->get_nmUser(),
-                'AKSES_USER' => $user->get_akses(),
-                'FOTO_USER' => $user->get_foto()
-            );
-            $this->_db->update($this->_table, $data, $where);
+        $data = array(
+            'NIP_USER' => $user->get_nip(),
+            'NM_USER' => $user->get_nmUser(),
+            'AKSES_USER' => $user->get_akses(),
+            'FOTO_USER' => $user->get_foto()
+        );
+        $this->_db->update($this->_table, $data, $where);
     }
-    
-    public function login($username,$password){
-        $sql = "SELECT * FROM ".$this->_table." WHERE NM_USER='".$username."' AND PASS_USER='".$password."'";
+
+    public function login($username, $password) {
+        $sql = "SELECT * FROM " . $this->_table . " WHERE NM_USER='" . $username . "' AND PASS_USER='" . $password . "'";
         $result = $this->_db->select($sql);
         $role = 0;
         $return = array();
-        foreach ($result as $v){
+        foreach ($result as $v) {
             $role = $v['AKSES_USER'];
             $kd = $v['NM_USER'];
         }
@@ -141,10 +141,59 @@ class User {
         $return[] = $kd;
         return $return;
     }
-    public function delUser($id) {
-        $where = "KD_USER=" . $id;
 
-        $this->_db->delete($this->_table, $where);
+    public function upload_foto() {
+
+        $allowedExts = array("pdf", "jpg", "jpeg", "png");
+
+        $ext = explode('.', $_FILES['upload']['name']);
+        $extension = $ext[count($ext) - 1];
+
+        if (in_array($extension, $allowedExts)) {
+
+            if ($_FILES["upload"]["error"] > 0) {
+                echo "Return Code: " . $_FILES["upload"]["error"] . "<br />";
+            } else {
+//                echo "Upload: " . $_FILES["upload"]["name"] . "<br />";
+//                echo "Type: " . $_FILES["upload"]["type"] . "<br />";
+//                echo "Size: " . ($_FILES["upload"]["size"] / 1024) . " Kb<br />";
+//                echo "Temp file: " . $_FILES["upload"]["tmp_name"] . "<br />";
+
+                if (file_exists("files/" . $_SESSION["nrp"] . "/" . $_FILES["upload"]["name"])) {
+                    echo $_FILES["upload"]["name"] . " file telah ada. ";
+                } else {
+                    // Create directory if it does not exist
+//                    if (!is_dir("files/" . $_SESSION["nrp"] . "/")) {
+//                        mkdir("files/" . $_SESSION["nrp"] . "/");
+//                    }
+                    // Move the uploaded file
+                    move_uploaded_file($_FILES["upload"]["tmp_name"], "files/" . $_SESSION["nrp"] . "/" . $_FILES["upload"]["name"]);
+                }
+            }
+        } else {
+            echo "Invalid file";
+        }
+//        header('location:' . URL . 'file');
+    }
+
+    public function delUser($id) {
+        
+        $where = "KD_USER=" . $id;
+        
+        $sql = "SELECT * FROM ".$this->_table." WHERE ".$where."";
+        $result = $this->_db->select($sql);
+        
+        $pic = new User($registry);
+        foreach ($result as $value) {
+            $pic->set_id($value['KD_USER']);
+            $pic->set_akses($value['AKSES_USER']);        
+        }
+        
+        if ($pic->get_akses()=='1'){
+            
+        } else {
+            $this->_db->delete($this->_table, $where);
+        }       
     }
 
     public function get_id() {
